+++ Change_Log.md
@@ -3,7 +3,17 @@ Day 1 (11/30/2025):
 - Set up github repo (this created default README.md and .attributes files)
 - created a .vscode file with launch.json and settings.json files to run the code with "love ."
 - created conf.lua file to configure game settings for title bar and game window size
-- put test code in main.lua (allows you to press d to move a square to the side
+- put test code in main.lua (allows you to press d to move a square to the side)
 
 Day 2 (12/1/2025):
-- 
+- No work
+
+Day 3 (12/2/2025):
+- Added nodes for map (still incomplete)
+
+Day 4 (12/3/2025):
+- Fixed node names
+- Load node into main.lua
+- Navigation through nodes via entering the node ID number via main.lua
+- Set up an image (should work when i add image paths to all nodes), still needs pixelated
+-